+++ packages/realm/bindgen/src/templates/node-wrapper.ts
@@ -16,14 +16,14 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 import { strict as assert } from "assert";
-import { bindModel, Property } from "@realm/bindgen/bound-model";
+import { Property } from "@realm/bindgen/bound-model";
 import { TemplateContext } from "@realm/bindgen/context";
 
 import { doJsPasses } from "../js-passes";
 import { eslint } from "../eslint-formatter";
 
-export function generate({ spec: rawSpec, file }: TemplateContext): void {
-  const spec = doJsPasses(bindModel(rawSpec));
+export function generate({ rawSpec, spec: boundSpec, file }: TemplateContext): void {
+  const spec = doJsPasses(boundSpec);
   const reactLines = [];
   const nodeLines = [];
   function both(...content: string[]) {
@@ -163,6 +163,8 @@ export function generate({ spec: rawSpec, file }: TemplateContext): void {
     `;
 
     for (const method of cls.methods) {
+      if (!method.isOptedInTo) continue;
+
       // Eagerly bind the name once from the native module.
       const native = `_native_${method.id}`;
       both(`const ${native} = nativeModule.${method.id};`);