+++ packages/realm/bindgen/src/templates/node.ts
@@ -20,7 +20,6 @@ import { strict as assert } from "assert";
 import { TemplateContext } from "@realm/bindgen/context";
 import { CppVar, CppFunc, CppFuncProps, CppCtor, CppMethod, CppClass, CppDecls } from "@realm/bindgen/cpp";
 import {
-  bindModel,
   BoundSpec,
   Class,
   InstanceMethod,
@@ -630,7 +629,7 @@ class NodeCppDecls extends CppDecls {
               body: `
                     auto out = Napi::Object::New(${env});
                     ${struct.fields
-                      .filter((field) => !field.type.isFunction())
+                      .filter((field) => !field.type.isFunction() && field.isOptedInTo)
                       .map(
                         (field) =>
                           `out.Set("${field.jsName}", ${convertToNode(
@@ -671,6 +670,7 @@ class NodeCppDecls extends CppDecls {
                 auto obj = val.As<Napi::Object>();
                 auto out = ${struct.cppName}();
                 ${struct.fields
+                  .filter((field) => field.isOptedInTo)
                   .map(
                     (field) => `{
                         auto field = obj.Get("${field.jsName}");
@@ -720,6 +720,8 @@ class NodeCppDecls extends CppDecls {
       };
 
       for (const method of cls.methods) {
+        if (!method.isOptedInTo) continue;
+
         const argOffset = method.isStatic ? 0 : 1; // `this` takes arg 0 if not static
         const args = method.sig.args.map((a, i) => convertFromNode(this.addon, a.type, `info[${i + argOffset}]`));
 
@@ -896,13 +898,13 @@ class NodeCppDecls extends CppDecls {
   }
 }
 
-export function generate({ spec, file: makeFile }: TemplateContext): void {
+export function generate({ rawSpec, spec, file: makeFile }: TemplateContext): void {
   const out = makeFile("node_init.cpp", clangFormat);
 
   // HEADER
   out(`// This file is generated: Update the spec instead of editing this file directly`);
 
-  for (const header of spec.headers) {
+  for (const header of rawSpec.headers) {
     out(`#include <${header}>`);
   }
 
@@ -915,7 +917,7 @@ export function generate({ spec, file: makeFile }: TemplateContext): void {
       namespace {
     `);
 
-  new NodeCppDecls(doJsPasses(bindModel(spec))).outputDefsTo(out);
+  new NodeCppDecls(doJsPasses(spec)).outputDefsTo(out);
 
   out(`
         } // namespace